+++ zetaclient/chains/sui/observer/outbound.go
@@ -83,7 +83,10 @@ func (ob *Observer) VoteOutbound(ctx context.Context, cctx *cctypes.CrossChainTx
 	// used checkpoint instead of block height
 	checkpoint, err := strconv.ParseUint(tx.Checkpoint, 10, 64)
 	if err != nil {
-		return errors.Wrap(err, "unable to parse checkpoint")
+		return errors.Wrapf(err, "invalid checkpoint: %s", tx.Checkpoint)
+	}
+	if checkpoint == 0 {
+		return errors.New("checkpoint is zero")
 	}
 
 	// parse outbound event
@@ -217,7 +220,11 @@ func (ob *Observer) postVoteOutbound(ctx context.Context, msg *cctypes.MsgVoteOu
 		retryGasLimit = zetacore.PostVoteOutboundRevertGasLimit
 	}
 
-	logger := ob.Logger().Outbound.With().Str(logs.FieldTx, msg.ObservedOutboundHash).Logger()
+	logger := ob.Logger().
+		Outbound.With().
+		Uint64(logs.FieldNonce, msg.OutboundTssNonce).
+		Str(logs.FieldTx, msg.ObservedOutboundHash).
+		Logger()
 
 	_, _, err := ob.ZetaRepo().VoteOutbound(ctx, logger, gasLimit, retryGasLimit, msg)
 	return err